requirements.txt
gitpython
python-dotenv
psutil
requests
networkx
matplotlib
bs4
pyvis